libkmod: move openssl-related functions to separate file
Move openssl-related function to a separate file so it's easier to
isolate the dependency on each crypto library.

This will allow multiple implementations to co-exist in the same build.

As a nice side effect we fallback to the "dummy" implementation in the
(very likely) case that the openssl one fails.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>